Add atomic trade protocol
Atomic trade protocol is completed in one request -> response
interaction between taker and maker.
1. Taker sends a CreateAtomicTxRequest to maker. This includes all the
data needed to create and sign the atomic tx from maker's side.
1. Maker verifies all inputs and answers with CreateAtomicTxResponse
that includes an atomic tx with maker's inputs signed.
1. Taker verifies the inputs and that their own outputs are paid as
expected, signs and publishes the completed atomic tx.
1. Fin

There is only one TakerVerifies and one MakerVerifies task as the buy
and sell side are very similar, the main difference lies in whether the
actor is taker or maker.

It's currently not possible to verify BTC inputs, but bad inputs (such
as spent ones) would generate a tx that won't broadcast. No funds can be
lost as the outputs are verified on both sides.