Register dispute agents in apitest arbitrator daemon
This PR adds a new CoreDisputeAgentService for handling dispute agent
registration in arbitrator daemons.  API methods for dispute agent
registration in mainnet mode are out of scope of this PR;  these
changes are for supporting the auto-registration of a mediator and
refund agent in regtest/dao mode for apitest cases.

Some minor, unrelated changes are included.

 * Inject new CoreDisputeAgentService into CoreApi.

 * Change a DisputeAgentManager 'dispute agent saved' log statement
   to indicate type of agent being saved.

 * Add import static bisq.common.app.Version.VERSION to CoreApi, and
   add @SuppressWarnings("SameReturnValue") annotation to the getVersion
   method.

 * Fix comment in Scaffold class.