Fix codacy issue caused by unused class level var
Commit 40dd1a9 @Injects a new CoreDisputeAgentService into CoreApi, but
it was unused by CoreApi, resulting in a legit codacy issue: "Perhaps
'coreDisputeAgentService' could be replaced by a local variable."

This change adds two pass through methods to CoreApi that use
the new CoreDisputeAgentService.  It should fix the codacy problem.
These new pass through methods are not useful yet, and may be
removed in a future PR for registering dispute agents with the API,
but CoreDisputeAgentService is needed now for apitest cases.

The CoreDisputeAgentService instance adds a listener to the
P2PService that tells it when the p2p network is bootstrapped (upon
receiving an onDataReceived event), then it automatically registers
dispute agents (if the daemon is an arbitrator node running in
regtest/dao mode).