Create new ping API method
A daemon will return an int32 = 1 (pong) to the CLI if successful,
then an "api service is available" message will be displayed in the
CLI's console.  If a valid pong is not returned, an "api service is
not available" message is displayed in the console.

Experiments will be made to see if this implementation can be used
to check server availabilty for other requests.  This server ping
implementation checks for a non-null wallet balance before returning
the pong (1) response.  If the ping's wallet check throws an exception
to the calling API method, the calling API method should throw a gRPC
exception to the CLI instead of completing the operation, and the user
should see an "api service is not available" message.

The premise is that if the wallet balance is available, so is
the server, else the server is not avaiable.  This may be a simple
solution to the 'p2p not ready' problem affecting daemons.

A ping test was added to apitest/scripts/mainnet-test.sh.