Perform network, chain and wallet checks in api's ping method
This change adds a new CorePingService that does a series of checks
before returning a pong, indicating the server is ready to perform
trading functions.

	* Check bsqNode.isP2pNetworkReady == true
	* Check p2PService.isBootstrapped == true
	* Check daoStateService.isParseBlockChainComplete == true
	* Check p2PService.numConnectedPeers >= 4
	* Check walletsManager.areWalletsAvailable == true
	* Check balances.availableBalance != null

An exception with a reason msg is thrown if any check fails.  That
exception will be logged in the server, and the CLI user will
see an "api service is not available" message in the console.

A public isP2pNetworkReady() acessor method was added to BsqNode to
make the bsqNode.isP2pNetworkReady == true possible.