Remove methods in MapStoreService and do instead instance of check in AppendOnlyDataStoreService to call the new method only where it makes sense.
Remove the hack of adding a "magic" key for communicating the version. Use a new field in the request instead.

Remove the extra handling of seedNode as the impl. did not make any difference  as the Version passed in
`excludedKeys = getKeySetInBytes(this.appendOnlyDataStoreService.getMap("since " + Version.VERSION));` was same as in
`if (!finalFilter.equals(Version.VERSION)) ` thus never doing anything there and the result maps is the same as in the seed node branch.

Was not clear anyway what the motivation of a diff. handling is there? That seed nodes request always all data without filtering? Why?