Redesign save file task
Remove executor which got started at each save request, leading to many tasks in case that saveLater was called frequently, which is specially at startup the case.
Use a periodic timer instead, makes the model much more simple.
Add saveNow methods with executor as param so it can be written on synchronously on user thread as well.
Add completeHandlers to the saveNow method.
Add isInitialized to check if the storage is already initialized before calling queueUpForSave