Move SellerPublishesDepositTx after SellerSendsDepositTxAndDelayedPayoutTxMessage
This should help to avoid that the buyer does not get the delayed payout tx.
If a timeout at deposit tx broadcast would prevent that it get out, the peer has at least both transactions.
The timeout could be Bisq internal but the tx gets in reality published. With the previous code we would get interrupted at
SellerPublishesDepositTx and then the SellerSendsDepositTxAndDelayedPayoutTxMessage would never get executed.