Create 'placeoffer' method
New offer placement is moved out of 'createoffer' into a separate
method, to give users a chance to review new offer details before
broadcasting it.

When createoffer is called, the new offer is cached in CoreOffersService
for 5 minutes and details are displayed in the CLI console.  The user
can broadcast the cached offer by calling 'placeoffer offer-id security-deposit'.
The security-deposit parameter is a % as double, e.g., 0.15.