Fixed Tests
The 'verify(this.mockSeqNrPersistenceManager, never()).requestPersistence();' got removes as they have never been working as expected.

requestPersistence gets called in populateTestState which calls the requestPersistence (queueUpForSave earlier).
After testing with the old version I found out the reason why it "worked" before:
The queueUpForSave method was invoked with param 2000 ms but the check here used anyLong() so it was treated
as different method call thus it was not failing. When I changed anyLong() to 2000 it failed as well.
So I will deactivate the test now as it was never working as expected anyway.
if (!expectedSeqNrWrite)
   verify(this.mockSeqNrPersistenceManager, never()).requestPersistence();