Remove persist method in trade and call it from outside callers on TradeManager instead
We have on case where set the state from inside the trade via setConfirmedState. As the confidence has its listener inside the trade as well this cannot be easily changed. But as when confidence of deposit tx changes we get anyway a outside task triggered and at restart we apply the confidence a confidence anyway so persistence is not really needed. As we persist at shutdown as well we have it covered there as well.
The confidence handling inside the trade should be refactored at some point so that trade does not get overloaded that much with behaviour code. Should be much more a value object....