Remove --runtime-image option using separate JDK
Remove the --runtime-image jpackage option, which was packaging Java 11 in the resulting installers. Reason is that jpackage 15 does not properly package the Java 11 runtime, but contains necessary features for signing and notarizing the app. On the other hand, jpackage 14 does package the Java 11 runtime correctly in the resulting installers, but does not support proper signing and notarizing.