Make URLs in popup notices clickable
Fixes #4623

URLs in Javafx label controls are not clickable, so separate them out to
one or more clickable Hyperlinks displayed at the footer of the popup.

The approach taken for this is to embed the tag `[HYPERLINK:]` in the
message around a link that you want to be handled this way.  The links
can be anywhere within the popup message, in the same order that they
will be displayed.  Using the tag has the advantage of being optional,
and does not require new code to be written for each individual popup.

Refactor to be more flexible; allowing inline hyperlinks

Codacy nits.