Clean up mailbox messages for closed trades
Util for removing pending mailbox messages in case the
trade has been closed by the seller after confirming receipt
and a AckMessage as mailbox message will be sent by the
buyer once they go online. In that case the seller's trade
is closed already and the TradeProtocol is not executing
the message processing, thus the mailbox message would not
be removed. To ensure that in such cases (as well other
potential cases in failure scenarios) the mailbox message
gets removed from the network we use that util.