Don't allow trade start if BitcoinJ is not fully synced
Adds a check that chain height is within 3 blocks of the reported
height from bitcoin peers -> if not the user cannot take an offer
or have an existing offer taken.  It shows a message informing the
user that Bisq is not currently synced, advising them to do an
SPV resync if necessary.

Additionally under Settings/Network a field has been added
to show the chain height of Bisq vs the Peer group.

Added after discussion with chimp1984:
- Correct initialization of chainHeight property
- Rename "Latest BTC block height" display field for clarity
- Enforce chain sync rule for Take Offer scenario
- Enforce chain synch rule for Check offer availability scenario
- change method name to be clearer